rosidl_codegen::types

Trait FieldTypeExt

pub trait FieldTypeExt {

    // Required methods
    fn is_sequence(&self) -> bool;
    fn is_primitive(&self) -> bool;
    fn is_string(&self) -> bool;
    fn is_unbounded_string(&self) -> bool;
    fn is_bounded_string(&self) -> bool;
    fn is_wstring(&self) -> bool;
    fn is_unbounded_wstring(&self) -> bool;
    fn is_bounded_wstring(&self) -> bool;
    fn is_primitive_sequence(&self) -> bool;
    fn is_string_sequence(&self) -> bool;
    fn is_unbounded_string_sequence(&self) -> bool;
    fn is_bounded_string_sequence(&self) -> bool;
    fn is_array(&self) -> bool;
    fn is_large_array(&self) -> bool;
    fn is_primitive_array(&self) -> bool;
    fn is_string_array(&self) -> bool;
    fn is_unbounded_string_array(&self) -> bool;
    fn is_bounded_string_array(&self) -> bool;
    fn is_unbounded_wstring_array(&self) -> bool;
    fn is_bounded_wstring_array(&self) -> bool;
    fn is_unbounded_wstring_sequence(&self) -> bool;
    fn is_bounded_wstring_sequence(&self) -> bool;
    fn is_nested_array(&self) -> bool;
    fn is_bounded_sequence(&self) -> bool;

}

Extension trait for FieldType providing type checking methods

Required Methods

fn is_sequence(&self) -> bool

Check if this is a sequence (unbounded or bounded)

fn is_primitive(&self) -> bool

Check if this is a primitive (no conversion needed)

fn is_string(&self) -> bool

Check if this is a string type (String, BoundedString, WString, BoundedWString)

fn is_unbounded_string(&self) -> bool

Check if this is specifically an unbounded string

fn is_bounded_string(&self) -> bool

Check if this is specifically a bounded string

fn is_wstring(&self) -> bool

Check if this is a WString type (unbounded or bounded)

fn is_unbounded_wstring(&self) -> bool

Check if this is specifically an unbounded WString

fn is_bounded_wstring(&self) -> bool

Check if this is specifically a bounded WString

fn is_primitive_sequence(&self) -> bool

Check if this is a sequence of primitives (can be copied directly)

fn is_string_sequence(&self) -> bool

Check if this is a sequence of strings (any string type)

fn is_unbounded_string_sequence(&self) -> bool

Check if this is a sequence of unbounded strings

fn is_bounded_string_sequence(&self) -> bool

Check if this is a sequence of bounded strings

fn is_array(&self) -> bool

Check if this is an array (needs clone, not conversion)

fn is_large_array(&self) -> bool

Check if this is a large array (> 32 elements, needs big_array for serde)

fn is_primitive_array(&self) -> bool

Check if this is an array of primitives (can be cloned directly)

fn is_string_array(&self) -> bool

Check if this is an array of strings (any string type)

fn is_unbounded_string_array(&self) -> bool

Check if this is an array of unbounded strings

fn is_bounded_string_array(&self) -> bool

Check if this is an array of bounded strings

fn is_unbounded_wstring_array(&self) -> bool

Check if this is an array of unbounded wstrings

fn is_bounded_wstring_array(&self) -> bool

Check if this is an array of bounded wstrings

fn is_unbounded_wstring_sequence(&self) -> bool

Check if this is a sequence of unbounded wstrings

fn is_bounded_wstring_sequence(&self) -> bool

Check if this is a sequence of bounded wstrings

fn is_nested_array(&self) -> bool

Check if this is an array of nested messages (needs element conversion)

fn is_bounded_sequence(&self) -> bool

Check if this is a bounded sequence (vs unbounded)